Wip 384 keycloak integration (#235)
* Adding authentication dependency to endpoints with FreePass
authenticator and jwt authenticator. Client works with both methods.

* Updated admin methods to include user addition into keycloak.
Created lomas_demo_setup script to be run in a separate container
(instead of filling up the mongodb from the fastapi server).
Move configs models and merge them when possible.

* Add keycloak user management to lomas dashboard. Remove some dashboard
tests until we have a proper dev environment that spins a mongodb and a
keycloak service.

* Update notebooks and clean client.

* PEP 639: SPDX license

---------

Co-authored-by: Bruno Studer <bruno.studer@protonmail.com>